Fix incorrect types causing fp precision errors with movable brush entities, etc
float -> double. This fixes an odd bug where some func_buttons would get stuck
after blocking them long enough.

Thanks to Mario (Xonotic) for finding this.

+++ sv_phys.c
@@ -1983,7 +1983,7 @@ SV_Physics_Pusher
 static void SV_Physics_Pusher (prvm_edict_t *ent)
 {
 	prvm_prog_t *prog = SVVM_prog;
-	float thinktime, oldltime, movetime;
+	double thinktime, oldltime, movetime;
 
 	oldltime = PRVM_serveredictfloat(ent, ltime);
